Fetch real data from spreadsheet, refactor
